added error and not-found
