--- app/error.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import React from 'react';
-
-export default function Error({
-  error,
-  reset,
-}: {
-  error: Error & { digest?: string };
-  reset: () => void;
-}) {
-  React.useEffect(() => {
-    console.error(error);
-  }, [error]);
-
-  return <>ist es over fur mich?</>;
-}
